fix: reuse types
