+++ lib/rag/retrieval/status.ts
@@ -2,7 +2,7 @@
 
 import { prisma } from '@/lib/prisma';
 import { RAG_CONFIG } from '../config';
-import type { AttachmentStatus } from './status-helpers';
+import type { AttachmentStatus } from '@/types/rag';
 import { partitionByStatus, extractIds } from './status-helpers';
 
 export async function getAttachmentStatuses(
