+++ types/tools.ts
@@ -1,0 +1,21 @@
+export interface WebSearchSource {
+  title: string;
+  url: string;
+  domain: string;
+  snippet: string;
+  score: number;
+  position: number;
+}
+
+export interface WebSearchProgress {
+  status: 'searching' | 'found' | 'processing_sources' | 'completed';
+  message: string;
+  details?: {
+    query?: string;
+    resultsCount?: number;
+    responseTime?: number;
+    sources?: WebSearchSource[];
+    currentSource?: WebSearchSource;
+    processedCount?: number;
+  };
+}
